Factor out DescriptionPackage's path-related functions to PackageLocator protocol
This would make it easy for testing FrameworkComponentsCollector and FrameworkModuleMapGenerator.
